Split out a build-report plugin
